Moved default menu initialisation in a seperate file
