+++ src/brackets.js
@@ -95,6 +95,7 @@ define(function (require, exports, module) {
         ExtensionUtils          = require("utils/ExtensionUtils");
             
     // Load modules that self-register and just need to get included in the main project
+    require("command/DefaultMenus");
     require("document/ChangedDocumentTracker");
     require("editor/EditorCommandHandlers");
     require("view/ViewCommandHandlers");
